Allow users to specify an itemCallback function.
Calling code can pass an 'itemCallback' function,
which will be invoked with two arguments,
the original file and the RSS item data object.
This is a caller's last chance to monkey with the
item data before it's added to the RSS enclosure.

One use case is to set categories or a GUID,
but others are possible, like doing some sort of
surgery on item titles or descriptions.